feat(postgres): replace GitHub PAT UI with PostgreSQL connection UI and validation
