+++ internal-packages/workflow-designer-ui/src/editor/properties-panel/text-generation-node-properties-panel/tools/tool-provider/github/index.tsx
@@ -4,23 +4,39 @@ import { Input } from "@giselle-internal/ui/input";
 import { Select } from "@giselle-internal/ui/select";
 import type { TextGenerationNode } from "@giselle-sdk/data-type";
 import { useWorkflowDesigner } from "@giselle-sdk/giselle/react";
-import { CheckIcon, PlusIcon, Settings2Icon, TrashIcon } from "lucide-react";
+import {
+	CheckIcon,
+	MoveUpRightIcon,
+	PlusIcon,
+	Settings2Icon,
+	TrashIcon,
+} from "lucide-react";
 import { Checkbox } from "radix-ui";
 import { useCallback, useState } from "react";
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "../ui/tabs";
+import { Tabs, TabsContent, TabsList, TabsTrigger } from "../../ui/tabs";
 import {
 	ToolConfigurationDialog,
 	type ToolConfigurationDialogProps,
-} from "../ui/tool-configuration-dialog";
-import { validatePostgreSQLConnectionString } from "./postgres-validation";
+} from "../../ui/tool-configuration-dialog";
 import {
 	ToolProviderSecretTypeValue,
 	useToolProviderConnection,
-} from "./use-tool-provider-connection";
+} from "../use-tool-provider-connection";
 
-const secretTags = ["postgres-connection-string"];
+const secretTags = ["github-access-token"];
 
-export function PostgresToolConfigurationDialog({
+// GitHub token validation
+function isValidGitHubPAT(token: string): boolean {
+	// GitHub token formats:
+	// Classic PAT: ghp_ followed by 36 alphanumeric characters (total 40 chars)
+	// Fine-grained PAT: github_pat_ followed by 82 alphanumeric characters (total 93 chars)
+	// OAuth token: gho_ followed by 36 alphanumeric characters (total 40 chars)
+	return /^(ghp_[a-zA-Z0-9]{36}|github_pat_[a-zA-Z0-9_]{82}|gho_[a-zA-Z0-9]{36})$/.test(
+		token,
+	);
+}
+
+export function GitHubToolConfigurationDialog({
 	node,
 }: {
 	node: TextGenerationNode;
@@ -37,14 +53,17 @@ export function PostgresToolConfigurationDialog({
 		handleSubmit,
 	} = useToolProviderConnection({
 		secretTags,
-		toolKey: "postgres",
+		toolKey: "github",
 		node,
-		buildToolConfig: (secretId) => ({ tools: [], secretId }),
+		buildToolConfig: (secretId) => ({
+			tools: [],
+			auth: { type: "secret", secretId },
+		}),
 	});
 
 	if (!isConfigured) {
 		return (
-			<PostgresToolConnectionDialog
+			<GitHubToolConnectionDialog
 				open={presentDialog}
 				onOpenChange={setPresentDialog}
 				tabValue={tabValue}
@@ -58,7 +77,7 @@ export function PostgresToolConfigurationDialog({
 	}
 
 	return (
-		<PostgresToolConfigurationDialogInternal
+		<GitHubToolConfigurationDialogInternal
 			node={node}
 			open={presentDialog}
 			onOpenChange={setPresentDialog}
@@ -66,7 +85,7 @@ export function PostgresToolConfigurationDialog({
 	);
 }
 
-function PostgresToolConnectionDialog({
+function GitHubToolConnectionDialog({
 	open,
 	onOpenChange,
 	tabValue,
@@ -83,26 +102,36 @@ function PostgresToolConnectionDialog({
 	secrets: { id: string; label: string }[] | undefined;
 	onSubmit: React.FormEventHandler<HTMLFormElement>;
 }) {
-	const [connectionString, setConnectionString] = useState("");
-	const [validationError, setValidationError] = useState<string | null>(null);
+	const [tokenValue, setTokenValue] = useState("");
+	const [tokenError, setTokenError] = useState<string | null>(null);
 
-	const handleConnectionStringChange = (value: string) => {
-		setConnectionString(value);
-		const validation = validatePostgreSQLConnectionString(value);
-		if (validation.isValid) {
-			setValidationError(null);
+	const handleTokenChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+		const value = e.target.value;
+		setTokenValue(value);
+
+		if (value && !isValidGitHubPAT(value)) {
+			setTokenError(
+				"Invalid token format. GitHub tokens should start with ghp_ (classic), github_pat_ (fine-grained), or gho_ (OAuth)",
+			);
 		} else {
-			setValidationError(validation.error);
+			setTokenError(null);
 		}
+	};
+
+	const handleFormSubmit = (e: React.FormEvent<HTMLFormElement>) => {
+		if (tabValue === "create" && tokenValue && !isValidGitHubPAT(tokenValue)) {
+			e.preventDefault();
+			return;
+		}
+		onSubmit(e);
 	};
 
 	return (
 		<ToolConfigurationDialog
-			title="Connect to PostgreSQL"
-			description="How would you like to connect to your database?"
-			onSubmit={onSubmit}
+			title="Connect to GitHub"
+			description="How would you like to add your Personal Access Token (PAT)?"
+			onSubmit={handleFormSubmit}
 			submitting={isPending}
-			disabled={tabValue === "create" && !!validationError}
 			trigger={
 				<Button type="button" leftIcon={<PlusIcon data-dialog-trigger-icon />}>
 					Connect
@@ -118,8 +147,8 @@ function PostgresToolConnectionDialog({
 				}
 			>
 				<TabsList className="mb-[12px]">
-					<TabsTrigger value="create">Paste Connection String</TabsTrigger>
-					<TabsTrigger value="select">Use Saved Connection String</TabsTrigger>
+					<TabsTrigger value="create">Paste New Token</TabsTrigger>
+					<TabsTrigger value="select">Use Saved Token</TabsTrigger>
 				</TabsList>
 				<TabsContent value="create">
 					<Input
@@ -130,19 +159,29 @@ function PostgresToolConnectionDialog({
 					<div className="flex flex-col gap-[12px]">
 						<fieldset className="flex flex-col">
 							<label htmlFor="label" className="text-text text-[13px] mb-[2px]">
-								Connection Name
+								Token Name
 							</label>
 							<Input type="text" id="label" name="label" />
 							<p className="text-[11px] text-text-muted px-[4px] mt-[1px]">
-								Give this connection a short name (e.g. “Prod-DB”). You’ll use
-								it when linking other nodes.
+								Give this token a short name (e.g. “Prod-bot”). You’ll use it
+								when linking other nodes.
 							</p>
 						</fieldset>
 						<fieldset className="flex flex-col">
 							<div className="flex justify-between mb-[2px]">
 								<label htmlFor="pat" className="text-text text-[13px]">
-									Connection String
+									Personal Access Token (PAT)
 								</label>
+								<a
+									href="https://github.com/settings/personal-access-tokens"
+									className="flex items-center gap-[4px] text-[13px] text-text-muted hover:bg-ghost-element-hover transition-colors px-[4px] rounded-[2px]"
+									target="_blank"
+									rel="noreferrer"
+									tabIndex={-1}
+								>
+									<span>GitHub</span>
+									<MoveUpRightIcon className="size-[13px]" />
+								</a>
 							</div>
 							<Input
 								type="password"
@@ -151,18 +190,23 @@ function PostgresToolConnectionDialog({
 								data-lpignore="true"
 								id="pat"
 								name="value"
-								value={connectionString}
-								onChange={(e) => handleConnectionStringChange(e.target.value)}
-								className={validationError ? "border-red-500" : ""}
+								value={tokenValue}
+								onChange={handleTokenChange}
+								aria-invalid={!!tokenError}
+								aria-describedby={tokenError ? "pat-error" : undefined}
 							/>
-							{validationError ? (
-								<p className="text-[11px] text-red-500 px-[4px] mt-[1px]">
-									{validationError}
+							{tokenError ? (
+								<p
+									id="pat-error"
+									className="text-[11px] text-red-600 px-[4px] mt-[1px]"
+									role="alert"
+								>
+									{tokenError}
 								</p>
 							) : (
 								<p className="text-[11px] text-text-muted px-[4px] mt-[1px]">
-									We'll encrypt the connection string with authenticated
-									encryption before saving it.
+									We'll encrypt the token with authenticated encryption before
+									saving it.
 								</p>
 							)}
 						</fieldset>
@@ -172,18 +216,18 @@ function PostgresToolConnectionDialog({
 					{isLoading ? (
 						<p>Loading...</p>
 					) : (secrets ?? []).length < 1 ? (
-						<EmptyState description="No saved connection strings.">
+						<EmptyState description="No saved tokens.">
 							<Button
 								onClick={() => setTabValue("create")}
 								leftIcon={<PlusIcon />}
 							>
-								Add a Connection String
+								Add a Token
 							</Button>
 						</EmptyState>
 					) : (
 						<>
 							<p className="text-[11px] text-text-muted my-[4px]">
-								Pick one of your encrypted connection strings.
+								Pick one of your encrypted tokens to connect.
 							</p>
 							<Input
 								type="hidden"
@@ -195,12 +239,12 @@ function PostgresToolConnectionDialog({
 									htmlFor="label"
 									className="text-text text-[13px] mb-[2px]"
 								>
-									Select a Saved Connection String
+									Select a Saved Token
 								</label>
 								<div>
 									<Select
 										name="secretId"
-										placeholder="Choose a connection string… "
+										placeholder="Choose a token… "
 										options={secrets?.map((s) => ({ ...s, value: s.id })) ?? []}
 										renderOption={(option) => option.label}
 										widthClassName="w-[180px]"
@@ -215,18 +259,60 @@ function PostgresToolConnectionDialog({
 	);
 }
 
-const postgresToolCatalog = [
+const githubToolCatalog = [
 	{
-		label: "Schema",
-		tools: ["getTableStructure"],
+		label: "Repository",
+		tools: ["getFileContents", "listBranches"],
 	},
 	{
-		label: "Query",
-		tools: ["query"],
+		label: "Issues",
+		tools: [
+			"createIssue",
+			"getIssue",
+			"listIssues",
+			"updateIssue",
+			"addIssueComment",
+			"getIssueComments",
+		],
 	},
+	{
+		label: "Pull Requests",
+		tools: [
+			"createPullRequest",
+			"getPullRequest",
+			"updatePullRequest",
+			"listPullRequests",
+			"getPullRequestComments",
+			"getPullRequestFiles",
+			"getPullRequestReviews",
+			"getPullRequestStatus",
+			"createPullRequestReview",
+			"addPullRequestReviewComment",
+			"mergePullRequest",
+			"updatePullRequestBranch",
+		],
+	},
+	{
+		label: "Code Management",
+		tools: ["createBranch", "createOrUpdateFile", "getCommit", "listCommits"],
+	},
+	{
+		label: "Search",
+		tools: [
+			"searchCode",
+			"searchIssues",
+			"searchPullRequests",
+			"searchRepositories",
+			"searchUsers",
+		],
+	},
+	// {
+	// 	label: "User",
+	// 	tools: ["getMe"],
+	// },
 ];
 
-function PostgresToolConfigurationDialogInternal({
+function GitHubToolConfigurationDialogInternal({
 	node,
 	open,
 	onOpenChange,
@@ -240,7 +326,7 @@ function PostgresToolConfigurationDialogInternal({
 	>(
 		(e) => {
 			e.preventDefault();
-			if (node.content.tools?.postgres === undefined) {
+			if (node.content.tools?.github === undefined) {
 				return;
 			}
 			const formData = new FormData(e.currentTarget);
@@ -252,8 +338,8 @@ function PostgresToolConfigurationDialogInternal({
 				...node.content,
 				tools: {
 					...node.content.tools,
-					postgres: {
-						...node.content.tools.postgres,
+					github: {
+						...node.content.tools.github,
 						tools,
 					},
 				},
@@ -265,8 +351,8 @@ function PostgresToolConfigurationDialogInternal({
 
 	return (
 		<ToolConfigurationDialog
-			title="PostgreSQL Configuration"
-			description="Select the PostgreSQL tools you want to enable."
+			title="GitHub Configuration"
+			description="Select the GitHub tools you want to enable."
 			onSubmit={updateAvailableTools}
 			submitting={false}
 			trigger={
@@ -284,7 +370,7 @@ function PostgresToolConfigurationDialogInternal({
 				<div className="flex justify-between items-center border border-border rounded-[4px] px-[6px] py-[3px] text-[13px] mb-[16px]">
 					<div className="flex gap-[6px] items-center">
 						<CheckIcon className="size-[14px] text-green-900" />
-						Connection configured.
+						Token configured.
 					</div>
 					<Button
 						type="button"
@@ -293,7 +379,7 @@ function PostgresToolConfigurationDialogInternal({
 								...node.content,
 								tools: {
 									...node.content.tools,
-									postgres: undefined,
+									github: undefined,
 								},
 							});
 						}}
@@ -304,7 +390,7 @@ function PostgresToolConfigurationDialogInternal({
 					</Button>
 				</div>
 				<div className="flex flex-col gap-6">
-					{postgresToolCatalog.map((category) => (
+					{githubToolCatalog.map((category) => (
 						<div key={category.label} className="flex flex-col gap-2">
 							<div className="text-[13px] font-medium text-text">
 								{category.label}
@@ -316,20 +402,38 @@ function PostgresToolConfigurationDialogInternal({
 										className="flex items-center justify-between p-3 hover:bg-black-800/30 cursor-pointer transition-colors"
 										htmlFor={tool}
 									>
-										<Checkbox.Root
-											className="group appearance-none size-[18px] rounded border flex items-center justify-center transition-colors outline-none data-[state=checked]:border-success data-[state=checked]:bg-success"
-											value={tool}
-											id={tool}
-											defaultChecked={node.content.tools?.postgres?.tools.includes(
-												tool,
-											)}
-											name="tools"
+										<div className="flex items-center flex-1">
+											<Checkbox.Root
+												className="group appearance-none size-[18px] rounded border flex items-center justify-center transition-colors outline-none data-[state=checked]:border-success data-[state=checked]:bg-success"
+												value={tool}
+												id={tool}
+												defaultChecked={
+													!!node.content.tools?.github?.tools?.includes(tool)
+												}
+												name="tools"
+											>
+												<Checkbox.Indicator className="text-background">
+													<CheckIcon className="size-[16px]" />
+												</Checkbox.Indicator>
+											</Checkbox.Root>
+											<p className="text-sm text-text flex-1 pl-[8px]">
+												{tool}
+											</p>
+										</div>
+										<a
+											href={`https://docs.giselles.ai/glossary/github-tools#${encodeURIComponent(
+												tool.toLowerCase(),
+											)}`}
+											target="_blank"
+											rel="noopener"
+											aria-label={`Open docs for ${tool}`}
+											className="flex items-center gap-[4px] text-[13px] text-text-muted hover:bg-ghost-element-hover transition-colors px-[4px] rounded-[2px] ml-2"
+											onMouseDown={(e) => e.stopPropagation()}
+											onClick={(e) => e.stopPropagation()}
 										>
-											<Checkbox.Indicator className="text-background">
-												<CheckIcon className="size-[16px]" />
-											</Checkbox.Indicator>
-										</Checkbox.Root>
-										<p className="text-sm text-text flex-1 pl-[8px]">{tool}</p>
+											<span>Docs</span>
+											<MoveUpRightIcon className="size-[13px]" />
+										</a>
 									</label>
 								))}
 							</div>
